Improve determine stem function and add specs
