+++ packages/yoastseo/spec/morphology/dutch/getStemWordsWithTAndDEndingSpecs.js
@@ -48,4 +48,12 @@ describe( "Creates stem from words with ambiguous endings", () => {
 	it( "Returns undefined if a word is not matched in any of the checks", () => {
 		expect( generateCorrectStemWithTAndDEnding( morphologyDataNL.stemming, "poolt" ) ).toEqual( null );
 	} );
+	it( "Creates the stem of a word which is in the list of words ending in -den that should only have -en stemmed.", () => {
+		expect( generateCorrectStemWithTAndDEnding( morphologyDataNL.stemming, "sportend" ) ).toEqual(
+			"sport" );
+	} );
+	it( "Creates the stem of a word which is in the list of words ending in -den that should only have -en stemmed.", () => {
+		expect( generateCorrectStemWithTAndDEnding( morphologyDataNL.stemming, "hoofden" ) ).toEqual(
+			"hoofd" );
+	} );
 } );
