+++ packages/yoastseo/src/morphology/dutch/stemTOrDFromEndOfWord.js
@@ -2,8 +2,6 @@ import { checkIfWordEndingIsOnExceptionList, checkExceptionListWithTwoStems } fr
 import { detectAndStemRegularParticiple } from "./detectAndStemRegularParticiple";
 import { generateCorrectStemWithTAndDEnding } from "./getStemWordsWithTAndDEnding";
 import checkExceptionsWithFullForms from "../morphoHelpers/checkExceptionsWithFullForms";
-import { removeSuffixFromFullForm } from "../morphoHelpers/stemHelpers";
-import nonParticiples from "../../researches/dutch/passiveVoice/nonParticiples.js";
 
 /**
  * If the word ending in -t/-d was not matched in any of the checks for whether -t/-d should be stemmed or not, and if it
@@ -18,26 +16,18 @@ import nonParticiples from "../../researches/dutch/passiveVoice/nonParticiples.j
  *
  * @returns {?string}				    The stemmed word or null if the -t/-d should not be stemmed.
  */
-export function createSecondStemWithoutAmbiguousEnding( morphologyDataNL, stemmedWord, word ) {
-	const ambiguousEndings = morphologyDataNL.stemming.stemExceptions.ambiguousTAndDEndings.tAndDEndings;
-	for ( const ending of ambiguousEndings ) {
-		// Check if the stem checked does not end in t/d. If it does not, return the stem.
-		if ( ! stemmedWord.endsWith( ending ) ) {
-			return stemmedWord;
-		}
-	}
-
-	// If the stem checked ends in t/d, run the checks below. If one of the conditions returns true, return the stem.
+export function stemTOrDFromEndOfWord( morphologyDataNL, stemmedWord, word ) {
+	// Run the checks below. If one of the conditions returns true, return the stem.
 	if ( detectAndStemRegularParticiple( morphologyDataNL, word ) ||
 		 generateCorrectStemWithTAndDEnding( morphologyDataNL.stemming, word ) ||
 		 checkIfWordEndingIsOnExceptionList( word, morphologyDataNL.stemming.stemExceptions.wordsNotToBeStemmedExceptions ) ||
-		 checkIfWordEndingIsOnExceptionList( word, morphologyDataNL.stemming.stemExceptions.adjectivesEndInRD ) ||
+		 checkIfWordEndingIsOnExceptionList( stemmedWord, morphologyDataNL.stemming.stemExceptions.adjectivesEndInRD ) ||
 		 checkExceptionsWithFullForms( morphologyDataNL.stemming.stemmingExceptionStemsWithFullForms, word ) ||
-		 removeSuffixFromFullForm( morphologyDataNL.stemming.stemExceptions.removeSuffixFromFullForms, word ) ||
-		 word.endsWith( "heid" ) || word.endsWith( "teit" ) || word.endsWith( "tijd" ) || nonParticiples().includes( word ) ) {
-		return stemmedWord;
+		 stemmedWord.endsWith( "heid" ) ) {
+		return null;
 	}
-	// If none of the conditions above true, stem the t/d from the word.
+
+	// If none of the conditions above is true, stem the t/d from the word.
 	stemmedWord = stemmedWord.slice( 0, -1 );
 	// Check if after t/d deletion, the word is in noun exception list.
 	const nounExceptionList = morphologyDataNL.nouns.exceptions.nounExceptionWithTwoStems;
